backend/requirements.txt
kafka-python
influxdb-client
python-dotenv